+++ include/ncei/models/cdo/data_type.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <nlohmann/json_fwd.hpp>
 #include <string>
 
 namespace ncei {
@@ -12,7 +11,5 @@ struct DataType {
 	std::string min_date;
 	std::string max_date;
 };
-
-void from_json(const nlohmann::json& j, DataType& d);
 
 } // namespace ncei